First major unit commit
This is a first pass at the unit code. Not only the scene and classes, but also a config driven apporach. Units are defined by static data only, and so must be created as a tres file. We also just load the types directly, and optionally define them in the civilization config. This does have the ui, but not yet tied to mouse clicks etc.